+++ repl-plots/src/main/scala/com/cibo/evilplot/displayPlot.scala
@@ -36,15 +36,22 @@ import com.cibo.evilplot.plot.Plot
 import javax.swing.{JFileChooser, JFrame, JPanel}
 import java.awt.event.{ActionEvent, ActionListener, ComponentAdapter, ComponentEvent}
 import java.io.File
+import java.util.prefs.Preferences
 
 import com.cibo.evilplot.geometry.{Drawable, Extent}
 import com.cibo.evilplot.plot.aesthetics.Theme
 import javax.swing.filechooser.FileNameExtensionFilter
 
+/** Helper utilities for displaying EvilPlot plots in their own window. The window offers
+  * dynamic resizing for plots and the ability to save the plot as a PNG.
+  *
+  * This utility is published in the `evilplot-repl` artifact, released alongside the
+  * regular `evilplot-jvm` artifact. It will need to be added to your build to use it.
+  */
 object displayPlot {
 
   private class DrawablePanel extends JPanel {
-    var drawable: Option[Drawable] = None
+    var drawable: Option[Drawable] = None // scalastyle:ignore
 
     def setDrawable(drawnPlot: Drawable): Unit = {
       drawable = Some(drawnPlot)
@@ -74,7 +81,7 @@ object displayPlot {
       val actionListener = new ActionListener {
         def actionPerformed(e: ActionEvent) = {
           val selectFile = new JFileChooser()
-          selectFile.setCurrentDirectory(null) //scalastyle:ignore
+          selectFile.setCurrentDirectory(loadLastSaveDir()) //scalastyle:ignore
           selectFile.setFileFilter(new FileNameExtensionFilter("png", "png"))
           val savedFile: Int = selectFile.showSaveDialog(panel)
           if (savedFile == JFileChooser.APPROVE_OPTION) {
@@ -90,6 +97,22 @@ object displayPlot {
       save.addActionListener(actionListener)
       menubar.add(save)
       setJMenuBar(menubar)
+    }
+
+    private def lastSaveDirPref = "lastSaveDir"
+
+    private def updateLastSaveDir(file: File): Unit = {
+      prefs.put(lastSaveDirPref, file.getParent)
+    }
+
+    private def loadLastSaveDir(): File = {
+      Option(prefs.get(lastSaveDirPref, null))
+        .map(new File(_))
+        .getOrElse(null)
+    }
+
+    private def prefs: Preferences = {
+      Preferences.userNodeForPackage(getClass).node("displayPlot")
     }
 
     private def init()(implicit theme: Theme): Unit = {
@@ -129,15 +152,24 @@ object displayPlot {
         case Right(d) => d.write(result)
         case Left(p)  => p.render(getPlotExtent).scaled(0.25, 0.25).write(result)
       }
+
+      updateLastSaveDir(result)
     }
 
   }
 
+  /** Display a plot in a JFrame. Passing in a plot makes the window resizable.
+    * @param plot The plot to render.
+    * @param theme The plot's theme. */
   def apply(plot: Plot)(implicit theme: Theme): Unit = {
     JFrame.setDefaultLookAndFeelDecorated(true)
     new DrawableFrame(Left(plot))
   }
 
+  /** Display any Drawable in a JFrame. Resizing the window does not affect the size of
+    * the rendered image.
+    * @param drawnPlot the Drawable to show in the window.
+    */
   def apply(drawnPlot: Drawable)(implicit theme: Theme): Unit = {
     JFrame.setDefaultLookAndFeelDecorated(true)
     new DrawableFrame(Right(drawnPlot))