+++ shared/src/main/scala/com/cibo/evilplot/plot/components/Axes.scala
@@ -41,7 +41,6 @@ object Axes {
   private sealed trait AxisPlotComponent extends PlotComponent {
     final override val repeated: Boolean = true
 
-    val discrete: Boolean
     val tickRenderer: TickRenderer
 
     def getDescriptor(plot: Plot, fixed: Boolean): AxisDescriptor
@@ -51,7 +50,6 @@ object Axes {
   }
 
   private sealed trait ContinuousAxis {
-    final val discrete: Boolean = false
     val tickCount: Int
     val tickCountRange: Option[Seq[Int]]
     val labelFormatter: Option[Double => String] = None
@@ -66,16 +64,15 @@ object Axes {
   }
 
   private sealed trait DiscreteAxis {
-    final val discrete: Boolean = true
     val labels: Seq[(String, Double)]
     def getDescriptor(plot: Plot, fixed: Boolean): AxisDescriptor = DiscreteAxisDescriptor(labels)
   }
 
   private sealed trait ArbitraryAxisPlotComponent extends AxisPlotComponent {
-    val fixedBounds: Boolean
+    val align: Double = 0
 
     override def size(plot: Plot): Extent = {
-      val extents = ticks(getDescriptor(plot, fixedBounds)).map(_.extent)
+      val extents = ticks(getDescriptor(plot, fixed = true)).map(_.extent)
       position match {
         case Position.Left | Position.Right => extents.maxBy(_.width)
         case Position.Bottom | Position.Top => extents.maxBy(_.height)
@@ -86,7 +83,7 @@ object Axes {
     def bounds(plot: Plot): Bounds
 
     def render(plot: Plot, extent: Extent)(implicit theme: Theme): Drawable = {
-      val descriptor = getDescriptor(plot, fixedBounds)
+      val descriptor = getDescriptor(plot, fixed = true)
       val scale = position match {
         case Position.Left | Position.Right => extent.height / descriptor.axisBounds.range
         case Position.Bottom | Position.Top => extent.width / descriptor.axisBounds.range
@@ -95,8 +92,8 @@ object Axes {
       val ts = ticks(descriptor)
       val maxWidth = ts.maxBy(_.extent.width).extent.width
       val maxHeight = ts.maxBy(_.extent.height).extent.height
-      // Move the tick to the center of the range for discrete axes.
-      val offset = (if (discrete) scale / 2 else 0) - scale * descriptor.axisBounds.min //TODO band scaling
+      // Move the tick to a position within the range for discrete axes.
+      val offset = scale * align - scale * descriptor.axisBounds.min
       position match {
         case Position.Left | Position.Right =>
           val drawable = ts
@@ -138,8 +135,7 @@ object Axes {
     tickCount: Int,
     tickRenderer: TickRenderer,
     override val labelFormatter: Option[Double => String],
-    tickCountRange: Option[Seq[Int]],
-    fixedBounds: Boolean = true
+    tickCountRange: Option[Seq[Int]]
   ) extends ArbitraryAxisPlotComponent
     with ContinuousAxis {
     override def bounds(plot: Plot): Bounds = boundsFn(plot)
@@ -149,10 +145,10 @@ object Axes {
     boundsFn: Plot => Bounds,
     override val position: Position,
     labels: Seq[(String, Double)],
-    tickRenderer: TickRenderer
+    tickRenderer: TickRenderer,
+    override val align: Double
   ) extends ArbitraryAxisPlotComponent
     with DiscreteAxis {
-    override val fixedBounds: Boolean = true
     override def bounds(plot: Plot): Bounds = boundsFn(plot)
   }
 
@@ -225,7 +221,6 @@ object Axes {
       * @param labelFormatter   Custom function to format tick labels.
       * @param tickCountRange   Allow searching over axis labels with this many ticks.
       * @param updatePlotBounds Set plot bounds to match the axis bounds.
-      * @param fixedBounds      Force ticks to match bounds.
       */
     def continuousAxis(
       boundsFn: Plot => Bounds,
@@ -234,8 +229,7 @@ object Axes {
       tickRenderer: Option[TickRenderer] = None,
       labelFormatter: Option[Double => String] = None,
       tickCountRange: Option[Seq[Int]] = None,
-      updatePlotBounds: Boolean = true,
-      fixedBounds: Boolean = true
+      updatePlotBounds: Boolean = true
     )(implicit theme: Theme): Plot = {
       val defaultRotation = if (position == Position.Bottom || position == Position.Top) {
         theme.elements.continuousXAxisLabelOrientation
@@ -254,15 +248,14 @@ object Axes {
             defaultRotation
           )),
         labelFormatter,
-        tickCountRange,
-        fixedBounds
+        tickCountRange
       )
       if (updatePlotBounds) {
         position match {
           case Position.Left | Position.Right =>
-            component +: plot.ybounds(component.getDescriptor(plot, fixedBounds).axisBounds)
+            component +: plot.ybounds(component.getDescriptor(plot, plot.yfixed).axisBounds)
           case Position.Bottom | Position.Top =>
-            component +: plot.xbounds(component.getDescriptor(plot, fixedBounds).axisBounds)
+            component +: plot.xbounds(component.getDescriptor(plot, plot.xfixed).axisBounds)
           case _ =>
             component +: plot
         }
@@ -277,15 +270,18 @@ object Axes {
       * @param position         The side of the plot to add the axis.
       * @param updatePlotBounds Set plot bounds to match the axis bounds.
       * @param tickRenderer     Function to draw a tick line/label.
+      * @param align            Where to align ticks as a proportion of their band, e.g. 0 = left, 0.5 = center.
       */
     def discreteAxis(
       labels: Seq[String],
       values: Seq[Double],
       position: Position,
       updatePlotBounds: Boolean = true,
-      tickRenderer: Option[TickRenderer] = None
+      tickRenderer: Option[TickRenderer] = None,
+      align: Double = 0.5
     )(implicit theme: Theme): Plot = {
       require(labels.lengthCompare(values.length) == 0)
+      require(0.0 <= align && align <= 1.0, "discreteAxis requires an align value from 0 to 1")
       val labelsAndValues = labels.zip(values)
       val (boundsFn, defaultRotation) = if (position == Position.Bottom || position == Position.Top) {
         ((plot: Plot) => plot.xbounds, theme.elements.categoricalXAxisLabelOrientation)
@@ -302,7 +298,8 @@ object Axes {
             theme.elements.tickLength,
             theme.elements.tickThickness,
             defaultRotation
-          ))
+          )),
+        align
       )
       if (updatePlotBounds) {
         position match {
@@ -340,8 +337,7 @@ object Axes {
         tickRenderer,
         labelFormatter,
         tickCountRange,
-        true,
-        plot.xfixed
+        true
       )
     }
 
@@ -408,8 +404,7 @@ object Axes {
         tickRenderer,
         labelFormatter,
         tickCountRange,
-        true,
-        plot.yfixed
+        true
       )
     }