+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Changelog
 
 ## [Changes since last release]
+### Added
+- Support for clustered box plots.
+- Option to display the data's mean as a dashed line on box plots.
+### Fixed
+- Bug in the spacing and sizing of bars in clustered bar charts.
 
+## [0.4.1] - 2018-07-26
 ### Added
 - Ability to partially update the bounds of a plot using `xbounds` and `ybounds`. Providing only `lower` or `upper` to these functions is now possible.
+- `BoxRenderer.tufte`, for an Edward Tufte style box plot.
+- New legend context builders.
+
+### Fixed
+- Bug resulting in misalignment of axes and gridlines in some faceted plots.
 
 ## [0.4.0] - 2018-07-18
 ### Added
@@ -60,7 +71,8 @@ correcting previously incorrect extent calculation.
 - Applying "bound buffers" to plots is no longer part of the default theme.
 - The default number of ticks on continuous axes has been decreased.
 
-[Changes since last release]: https://github.com/cibotech/evilplot/compare/v0.4.0...HEAD
+[Changes since last release]: https://github.com/cibotech/evilplot/compare/v0.4.1...HEAD
+[0.4.1]: https://github.com/cibotech/evilplot/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/cibotech/evilplot/compare/v0.3.4...v0.4.0
 [0.3.4]: https://github.com/cibotech/evilplot/compare/v0.3.3...v0.3.4
 [0.3.3]: https://github.com/cibotech/evilplot/compare/v0.3.2...v0.3.3